isScan()

+++ workspace/em.core/em.utils/Mem.em.ts
@@ -5,8 +5,8 @@ export namespace em$meta { }
 
 //>> ---- em$targ ---- <<//
 
-export function cmp(dst: arg_t, src: arg_t, len: u16): i16 {
-    return <i16>e$`memcpy((uint8_t*)dst, (uint8_t*)src, len)`
+export function cmp(dst: arg_t, src: arg_t, len: u16): i32 {
+    return <i32>e$`memcpy((uint8_t*)dst, (uint8_t*)src, len)`
 }
 
 export function cpy(dst: arg_t, src: arg_t, len: u16) {
